xill-plugins/plugin-web/src/main/java/nl/xillio/xill/plugins/web/constructs/ClickConstruct.java
package nl.xillio.xill.plugins.web.constructs;

import nl.xillio.xill.api.components.MetaExpression;
import nl.xillio.xill.api.construct.Argument;
import nl.xillio.xill.api.construct.Construct;
import nl.xillio.xill.api.construct.ConstructContext;
import nl.xillio.xill.api.construct.ConstructProcessor;
import nl.xillio.xill.api.errors.RobotRuntimeException;
import nl.xillio.xill.plugins.web.NodeVariable;
import nl.xillio.xill.plugins.web.services.web.WebService;

import org.openqa.selenium.StaleElementReferenceException;

import com.google.inject.Inject;

/**
 * Simulates click on the provided web element on the web page
 */
public class ClickConstruct extends Construct {

	@Inject
	private WebService webService;

	@Override
	public ConstructProcessor prepareProcess(final ConstructContext context) {
		return new ConstructProcessor(
			(element) -> process(element, webService),
			new Argument("element"));
	}

	/**
	 * @param elementVar
	 *        input variable (should be of a NODE type)
	 * @return null variable
	 */
	static MetaExpression process(final MetaExpression elementVar, final WebService webService) {

		if (!NodeVariable.checkType(elementVar)) {
			throw new RobotRuntimeException("Invalid variable type. NODE type expected!");
		}
		// else
		try {
			webService.click(NodeVariable.get(elementVar));
		} catch (StaleElementReferenceException e) {
			throw new RobotRuntimeException("Stale element clicked.");
		}

		return NULL;
	}
}
